Testing for nextConfig/withFaust
