+++ examples/next/getting-started/next.config.js
@@ -1,3 +1,6 @@
 const { withFaust } = require('@faustjs/next');
 
+/**
+ * @type {import('next').NextConfig}
+ **/
 module.exports = withFaust();
