--- packages/next/src/nextConfig.tsx
@@ -1,26 +1,0 @@
-import merge from 'lodash/merge';
-import { NextConfig } from 'next';
-
-export function withFaust(config: NextConfig): NextConfig {
-  const faustConfig: NextConfig = {
-    // eslint-disable-next-line @typescript-eslint/require-await
-    async redirects() {
-      return [
-        {
-          source: '/((?!preview$).*)',
-          has: [
-            {
-              type: 'query',
-              key: 'preview',
-              value: 'true',
-            },
-          ],
-          destination: '/preview',
-          permanent: false,
-        },
-      ];
-    },
-  };
-
-  return merge(config, faustConfig);
-}
